organized controllers so we don't have all endpoint functions in the same file
