+++ controllers/user_controller.go
@@ -5,11 +5,8 @@ import (
 	"github.com/ContainerMaintainers/MiniTwit-Golang/database"
 	"github.com/ContainerMaintainers/MiniTwit-Golang/infrastructure/entities"
 	"github.com/gin-gonic/gin"
-	"golang.org/x/crypto/bcrypt"
 	"log"
 	"net/http"
-	"net/url"
-	"strings"
 )
 
 var (
@@ -95,195 +92,6 @@ func usernameUnfollow(c *gin.Context) { // Adds the current user as follower of 
 		//c.String(200, fmt.Sprintf("You are no longer following \"%s\"", username)) // Had to make it 200 to satisfy tests
 		c.Redirect(http.StatusFound, "/"+username)
 	}
-}
-
-// login helper function
-func valid_login(c *gin.Context, message string, user int) bool {
-	if message == "Invalid username" {
-		log.Print("Bad request during " + c.Request.RequestURI + ": " + " Invalid Username")
-		c.HTML(http.StatusOK, "login.html", gin.H{
-			"error": "Invalid Username",
-		})
-		return false
-	} else if message == "Invalid password"  {
-		log.Print("Bad request during " + c.Request.RequestURI + ": " + " Invalid password")
-		c.HTML(http.StatusOK, "login.html", gin.H{
-			"error": "Invalid password",
-		})
-		return false
-	} else if message == "You are logged in!" {
-		c.HTML(http.StatusOK, "timeline.html", gin.H{
-			"title":     "My Timeline",
-			"flashes":	 message,
-			"user":      user,
-			"private":   true,
-			"user_page": true,
-			"messages":  GetMessages("myTimeline", user, 0),
-		})
-		return true
-	}
-	return true
-	
-}
-
-// ENDPOINT: POST /login
-func login_user(c *gin.Context) { //Logs the user in.
-
-	var body struct {
-		Username string `form:"username" json:"username"`
-		Password string `form:"password" json:"password"`
-	}
-
-	err := c.Bind(&body)
-	if err != nil {
-		log.Print("Ran into error when binding to context during " + c.Request.RequestURI + ": " + err.Error())
-		c.AbortWithStatus(400)
-		return
-	}
-
-	if _, err := getUserId(body.Username); err != nil {
-		// If invalid username
-		log.Print("Bad request during " + c.Request.RequestURI + ": Invalid username " + body.Username)
-		valid_login(c, "Invalid username", -1)
-	} else if _, err := checkPasswordHash(body.Username, body.Password); err != nil {
-		// if invalid password
-		log.Print("Ran into error during " + c.Request.RequestURI + ": " + err.Error() + ", invalid password")
-		valid_login(c, "Invalid password", -1)
-	} else {
-		if userID, err := getUserId(body.Username); err != nil {
-			log.Print("Ran into error during " + c.Request.RequestURI + ": " + err.Error())
-			user = -1
-		} else {
-			// set session user
-			user = int(userID)
-			valid_login(c, "You are logged in!", user)
-		}
-	}	
-}
-
-// ENDPOINT: GET /logout
-func logout_user(c *gin.Context) {
-	//clear session user
-	user = -1
-	
-	//c.String(200, "You were logged out")
-	c.Redirect(http.StatusFound, "/")
-}
-
-// ENDPOINT: GET /register
-func register(c *gin.Context) {
-	c.HTML(http.StatusOK, "register.html", gin.H{
-		"messages": "register page",
-	})
-}
-
-// ENDPOINT: GET /login
-func loginf(c *gin.Context) {
-	c.HTML(http.StatusOK, "login.html", gin.H{
-		"messages": "Login page",
-	})
-}
-
-// Registration Helper Function
-func ValidRegistration(c *gin.Context, username string, email string, password1 string, password2 string) bool {
-	
-	//error = ""
-	if password1 == "" || email == "" || username == "" {
-		log.Print("Bad request during " + c.Request.RequestURI + ": " + " Missing Field")
-		c.HTML(http.StatusOK, "register.html", gin.H{
-			"error": "All fields are required",
-		})
-		return false
-	} else if email == "" || !strings.Contains(email, "@") {
-		log.Print("Bad request during " + c.Request.RequestURI + ": " + " Invalid email")
-		c.HTML(http.StatusOK, "register.html", gin.H{
-			"error": "You have to enter a valid email address",
-		})
-		return false
-	} else if password1 != password2 {
-		log.Print("Bad request during " + c.Request.RequestURI + ": " + " Passwords do not match")
-		c.HTML(http.StatusOK, "register.html", gin.H{
-			"error": "The two passwords do not match",
-		})
-		return false
-	} else if _, err := getUserId(username); err == nil {
-		log.Print("Ran into error during " + c.Request.RequestURI + ": " + " Username already taken")
-		c.HTML(http.StatusOK, "register.html", gin.H{
-			"error": "The username is already taken",
-		})
-		return false
-	}
-	return true
-}
-
-// ENDPOINT: POST /register
-func register_user(c *gin.Context) {
-
-	var body struct {
-		Username  string `form:"username" json:"username"`
-		Email     string `form:"email" json:"email"`
-		Password  string `form:"password" json:"password"`
-		Password2 string `form:"password2" json:"password2"`
-	}
-
-	err := c.Bind(&body)
-	if err != nil {
-		log.Print("Ran into error when binding to context during " + c.Request.RequestURI + ": " + err.Error())
-		c.AbortWithStatus(400)
-		return
-	}
-
-	error := ""
-
-	if ValidRegistration(c, body.Username, body.Email, body.Password, body.Password2) {
-		user := entities.User{
-			Username: body.Username,
-			Email:    body.Email,
-			Password: Salt_pwd(body.Password),
-		}
-
-		err := database.DB.Create(&user).Error
-		if err != nil {
-			log.Print("Ran into error during " + c.Request.RequestURI + ": " + err.Error())
-			c.AbortWithStatus(500)
-			return
-		}
-
-		//c.String(200, "You were successfully registered and can login now")
-		location := url.URL{Path: "/login"}
-		c.Redirect(http.StatusFound, location.RequestURI())
-
-	} else {
-		c.String(400, error)
-	}
-
-}
-
-func Salt_pwd(pwd string) string {
-	hash, err := bcrypt.GenerateFromPassword([]byte(pwd), bcrypt.MinCost)
-	if err != nil {
-		log.Fatal("The given password could not be hashed. ", err)
-	}
-
-	return string(hash)
-}
-
-func checkPasswordHash(username string, enteredPW string) (bool, error) {
-	var user entities.User
-
-	if err := database.DB.Where("username = ?", username).First(&user).Error; err != nil {
-		return false, err
-	}
-
-	if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(enteredPW)); err != nil {
-		return false, err
-	}
-
-	if err := database.DB.Where("username = ? AND password = ?", username, user.Password).First(&user).Error; err != nil {
-		return false, err
-	}
-
-	return true, nil
 }
 
 func getUserId(username string) (uint, error) { // Convenience method to look up the id for a username.
